+++ src/types.ts
@@ -211,3 +211,10 @@ export interface PanPressableOptions {
      */
     longPressMaxDistance?: number;
 }
+
+export interface LayoutLinkOptions {
+    axis: LayoutLinkAxis;
+    mode?: 'auto' | 'manual';
+}
+
+export type LayoutLinkAxis = 'x' | 'y' | 'xy';